kernelci.config.db: rename from kernelci.config.data
Rename kernelci.config.data into kernelci.config.db to match
kernelci.db and avoid ambibuity with data files.  Update all the
places where kernelci.config.data was being used.
